Mudanca_Vencimento/requirements.txt
fastapi
pandas
uvicorn 
psycopg2
pyodbc
requests
python-dotenv